fix AppVeyor

+++ src/lug/Graphics/Vulkan/Pipeline.cpp
@@ -109,8 +109,8 @@ std::unique_ptr<Pipeline> Pipeline::createGraphicsPipeline(const Device* device,
     VkViewport viewport{
         viewport.x = 0.0f,
         viewport.y = 0.0f,
-        viewport.width = extent.width,
-        viewport.height = extent.height,
+        viewport.width = static_cast<float>(extent.width),
+        viewport.height = static_cast<float>(extent.height),
         viewport.minDepth = 0.0f,
         viewport.maxDepth = 1.0f,
     };
